LRCI-3761 Consistency

+++ modules/test/jenkins-results-parser/src/main/java/com/liferay/jenkins/results/parser/JenkinsResultsParserUtil.java
@@ -4445,14 +4445,14 @@ public class JenkinsResultsParserUtil {
 
 				String exceptionMessage = ioException.getMessage();
 
-				Long retryPeriodOverride = null;
+				Integer retryPeriodOverride = null;
 
 				if (exceptionMessage.matches(
 						".*HTTP response code\\: 403 .*") &&
 					(urlConnection != null)) {
 
 					try {
-						retryPeriodOverride = Long.parseLong(
+						retryPeriodOverride = Integer.parseInt(
 							urlConnection.getHeaderField("retry-after"));
 					}
 					catch (NumberFormatException numberFormatException) {
@@ -4481,7 +4481,7 @@ public class JenkinsResultsParserUtil {
 					if ((retryPeriodOverride == null) ||
 						(retryPeriodOverride == 0)) {
 
-						retryPeriodOverride = (long)retryPeriod;
+						retryPeriodOverride = retryPeriod;
 
 						for (int i = 0; i < retryCount; i++) {
 							retryPeriodOverride *= retryPeriodOverride;
